Setup basic layout
The basic layout of the application is set in the template files. At the
top layer the app contains the header at the top. And the below the
header, there is recipes and shopping-list components.

The recipes component have a layout where the left side shows the
recipe-list and the right side shows the recipe-details.

The shopping-list component contains the shopping-edit component at the
top and the list at the bottom.